==> 1. ETL_Siniestros_Viales.ipynb <==
# Siniestros Viales en la Ciudad Autónoma de Buenos Aires
### (Homicidos en Accidentes de Tráfico)

Con base en los datos alojados en la página de "Buenos Aires Data" se extraen los datasets de "homicidios" en Excel que contienen las víctimas mortales, lugar de los hechos y los actores involucrados.

Extracción, Transformación y Limpiza de los datos:

#### 1. Descarga de librería necesarias:

import pandas as pd                                                           # Para abrir y transformar los archivos
from urllib.request import urlretrieve                                        # Para leer y guardar los archivos originales
import openpyxl                                                               # Para manipular los archivos en excel

#### 2. Descargar archivo de "homicidios" y guardarlo en la carpeta "Datos_Originales":

# URL del archivo Excel
url_homicidios = 'https://cdn.buenosaires.gob.ar/datosabiertos/datasets/transporte-y-obras-publicas/victimas-siniestros-viales/homicidios.xlsx'

# Ruta local donde se guardará el archivo
Ubicación_homicidios = "Datos_Originales/Homicidios_CABA.xlsx"

# Descargar el archivo desde la URL y guardarlo localmente
urlretrieve(url_homicidios, Ubicación_homicidios)

# Crear un objeto ExcelFile para manejar la estructura del archivo Excel
wb_Homicidios = openpyxl.load_workbook(Ubicación_homicidios, data_only=True)

# Obtener la lista de nombres de hojas
sheet_names = wb_Homicidios.sheetnames

# Imprimir los nombres de las hojas disponibles
print("Hojas disponibles:", sheet_names)

A través de openpyxl, se crea un objeto "ExcelFile" del que se pueden obtener los nombres con el método "Sheetnames" y manipular eficientemente el archivo de excel.

#### 3. Crear Datafrome de la hoja de "Hechos" del archivo de excel "Homicidios":

# Crear un objeto ExcelFile para manejar el archivo Excel
xl = pd.ExcelFile(Ubicación_homicidios)

# crear dataframe
df_homicidios = xl.parse('HECHOS',skiprows=0)

# Reiniciar los índices después de la eliminación
df_homicidios.reset_index(drop=True, inplace=True)

#### 3.1. Revisar la calidad de la información por columnas:

df_homicidios.info()

Aparentemente solo los campos "Cruce" y "Altura" tienen registros nulos, pero se debe revisar más específicamente el contenido de las demás columnas.

A continuación se revisan los datos nulos de las columnas que se consideran relevantes:

print('Victima:',df_homicidios['VICTIMA'].unique())
print('--------------------------------------------')
print('Acusado:',df_homicidios['ACUSADO'].unique())
print('--------------------------------------------')
print('Comuna:',df_homicidios['COMUNA'].unique()) # Observamos comuna con valores de "0" lo cual no es real
print('--------------------------------------------')
print('pos x:',df_homicidios['pos x'].min()," hasta ",df_homicidios['pos x'].max()) # datos con "." que deben ser revisados
print('pos y:',df_homicidios['pos y'].min()," hasta ",df_homicidios['pos y'].max()) # datos con "." que deben ser revisados
print('--------------------------------------------')
print('Participantes:',df_homicidios['PARTICIPANTES'].unique()) #Se tiene un valor errado 'PEATON_MOTO-MOTO'

#### 3.1.1. Revisar columna de "COMUNA" con Valor "0":

comunas_con_valor_cero = df_homicidios[df_homicidios['COMUNA'] == 0]
comunas_con_valor_cero 

Revisar columnas con Autopista 25 de mayo:

# Filtrar las filas donde 'Dirección Normalizada' contiene 'AUTOPISTA 25'
Autopista25 = df_homicidios[df_homicidios['Dirección Normalizada'].str.contains('AUTOPISTA 25', case=False, na=False)]

# Mostrar el DataFrame resultante
Autopista25.tail()

debido a que 25 de Mayo de relaciona con varias comunas se imputará el "0" como "SD":

# Imputar "SD" en lugar de "0" en la columna 'COMUNA'
df_homicidios.loc[df_homicidios['COMUNA'] == 0, 'COMUNA'] = 'SD'

comunas_con_valor_cero = df_homicidios[df_homicidios['COMUNA'] == 'SD']
comunas_con_valor_cero 

#### 3.1.2. Revisar los datos de coordenadas:

# Filtrar las filas donde 'Dirección Normalizada' contiene 'AUTOPISTA 25'
Coordenada_X0 = df_homicidios[df_homicidios['pos x']=='.']

# Mostrar el DataFrame resultante
Coordenada_X0.tail()

Se imputarán las coordenadas basados en la comuna y tipo de calle, a continuación se realiza la imputación para el eje x:

# Filtrar las filas donde 'pos x' es igual a '.'
Coordenada_X0 = df_homicidios[df_homicidios['pos x'] == '.']

# Obtener un DataFrame de referencia con valores de 'pos x' que no son '.'
Referencia_X0 = df_homicidios[df_homicidios['pos x'] != '.']

# Imputar los valores de 'pos x' basándose en la comuna de la Referencia_X0
def imputar_pos_x(row):
    filtro = (Referencia_X0['COMUNA'] == row['COMUNA'])
    if filtro.any():
        return Referencia_X0[filtro]['pos x'].iloc[0]
    else:
        return row['pos x']

Coordenada_X0['pos x'] = Coordenada_X0.apply(imputar_pos_x, axis=1)

# Actualizar los valores en el DataFrame original
df_homicidios.loc[Coordenada_X0.index, 'pos x'] = Coordenada_X0['pos x']

a continuación se realiza la imputación para el eje y:

# Filtrar las filas donde 'pos y es igual a '.'
Coordenada_Y0 = df_homicidios[df_homicidios['pos y'] == '.']

# Obtener un DataFrame de referencia con valores de 'pos y' que no son '.'
Referencia_Y0 = df_homicidios[df_homicidios['pos y'] != '.']

# Imputar los valores de 'pos y' basándose en la comuna de la Referencia_Y0
def imputar_pos_y(row):
    filtro = (Referencia_Y0['COMUNA'] == row['COMUNA'])
    if filtro.any():
        return Referencia_Y0[filtro]['pos y'].iloc[0]
    else:
        return row['pos y']

Coordenada_Y0['pos y'] = Coordenada_Y0.apply(imputar_pos_y, axis=1)

# Actualizar los valores en el DataFrame original
df_homicidios.loc[Coordenada_Y0.index, 'pos y'] = Coordenada_Y0['pos y']

Cambiar los registros en coordenadas "." con "SD", son solo dos registros que se manterndrán porque suma en los indicadores, con el campo "SD" puede filtrarse (omitirse) estos datos desde el dashboard:

df_homicidios.loc[df_homicidios['pos y'] == '.', 'pos y'] = 'SD'
df_homicidios.loc[df_homicidios['pos x'] == '.', 'pos x'] = 'SD'

#### 3.1.3. Revisión de la dirección normalizada que presenta datos nulos:

# Filtrar las filas donde 'pos y es igual a '.'
dirección_SD = df_homicidios[df_homicidios['Dirección Normalizada'].isna()]
dirección_SD

df_homicidios['Dirección Normalizada'].nunique()

la dirección en su mayoría representa valores unicos y esto no serviría mucho en clasificación por zonas de los accidentes, así que se decide usar coordenadas y comunas como referecias. No se imputarán direcciones con lugar del hecho, así esto sea posible porque dirección será elimnado.

#### 3.1.4. Eliminar columna innecesarias y editar tipos de datos correspondientes:

# Lista de columnas a eliminar
columnas_a_eliminar = ['N_VICTIMAS', 'AAAA', 'MM', 'DD', 'HH', 'LUGAR_DEL_HECHO', 'Calle', 'Altura', 'Cruce', 'Dirección Normalizada', 'XY (CABA)', 'PARTICIPANTES']

# Eliminar las columnas especificadas
df_homicidios = df_homicidios.drop(columns=columnas_a_eliminar)

# Mostrar el DataFrame actualizado
df_homicidios

df_homicidios.info()

#### 3.2. Exportar a CSV en la carpeta de Datos_Limpios

df_homicidios.to_csv('Datos_limpios/Hechos_Homicidios_CABA.csv',index=False)

#### 4. Crear Datafrome de la hoja de "Víctimas" del archivo de excel "Homicidios":

# crear dataframe
df_victimas = xl.parse('VICTIMAS',skiprows=0)

# Reiniciar los índices después de la eliminación
df_victimas.reset_index(drop=True, inplace=True)

##### 4.1. Revisamos la calidad de los datos y datos nulos:

df_victimas.info()

# Contar los valores "SD" en todas las columnas
conteo_SD_por_columna = (df_victimas == "SD").sum()

# Mostrar el resultado
print(conteo_SD_por_columna)

#### 4.2. Elimar columnas innecesarias y cambiar tipo correspondiente en fecha:

# Lista de columnas a eliminar
columnas_vic_eliminar = ['AAAA', 'MM', 'DD', 'FECHA_FALLECIMIENTO']

# Eliminar las columnas especificadas
df_victimas = df_victimas.drop(columns=columnas_vic_eliminar)

# Formato de fecha
df_victimas['FECHA'] = pd.to_datetime(df_victimas['FECHA'], errors='coerce')

# Mostrar el DataFrame actualizado
df_victimas

df_victimas.info()

#### 4.2. Exportar archivo a la carpeta "Datos_Limpios":

df_victimas.to_csv('Datos_limpios/Victimas_Homicidios_CABA.csv',index=False)
