==> 2. ETL_Lesiones_No_Fatales.ipynb <==
# Siniestros Viales en la Ciudad Autónoma de Buenos Aires
### (Lesiones no fatales)

Con base en los datos alojados en la página de "Buenos Aires Data" se extraen los datasets de "homicidios" en Excel que contienen las víctimas mortales, lugar de los hechos y los actores involucrados.

Extracción, Transformación y Limpiza de los datos:

#### 1. Descarga de librería necesarias:

import pandas as pd                                                           # Para abrir y transformar los archivos
from urllib.request import urlretrieve                                        # Para leer y guardar los archivos originales
import openpyxl                                                               # Para manipular los archivos en excel
from pyproj import Proj, transform                                            # Para transformar coordenadas
import numpy as np                                                            # Para tratar nulos y realizar operaciones matemáticas
import warnings
warnings.filterwarnings("ignore")                                             # Ignorar todas las advertencias

#### 2. Descargar archivo de "homicidios" y guardarlo en la carpeta "Datos_Originales":

# URL del archivo Excel
url_lesiones = 'https://cdn.buenosaires.gob.ar/datosabiertos/datasets/transporte-y-obras-publicas/victimas-siniestros-viales/lesiones.xlsx'

# Ruta local donde se guardará el archivo
Ubicación_lesiones = "Datos_Originales/Lesiones_CABA.xlsx"

# Descargar el archivo desde la URL y guardarlo localmente
urlretrieve(url_lesiones, Ubicación_lesiones)

# Crear un objeto ExcelFile para manejar la estructura del archivo Excel
wd_lesiones = openpyxl.load_workbook(Ubicación_lesiones, data_only=True)

# Obtener la lista de nombres de hojas
sheet_names_lesiones = wd_lesiones.sheetnames

# Imprimir los nombres de las hojas disponibles
print("Hojas disponibles:", sheet_names_lesiones)

A través de openpyxl, se crea un objeto "ExcelFile" del que se pueden obtener los nombres con el método "Sheetnames" y manipular eficientemente el archivo de excel.

#### 3. Crear Datafrome de la hoja de "Hechos" del archivo de excel "Lesiones":

# Crear un objeto ExcelFile para manejar el archivo Excel
xl = pd.ExcelFile(Ubicación_lesiones)

# crear dataframe
df_lesiones = xl.parse('HECHOS',skiprows=0)

# Reiniciar los índices después de la eliminación
df_lesiones.reset_index(drop=True, inplace=True)

#### 3.1. Revisar la calidad de la información por columnas:

df_lesiones.info()

Aparentemente solo los campos "Cruce" y "Altura" tienen registros nulos, pero se debe revisar más específicamente el contenido de las demás columnas.

A continuación se revisan los datos nulos de las columnas que se consideran relevantes:

print('Victima:',df_lesiones['victima'].unique())  # Normalizar los datos en función del archivo de Homicidios
print('--------------------------------------------')   # Normalizar los datos en función del archivo de Homicidios
print('Acusado:',df_lesiones['acusado'].unique())
print('--------------------------------------------')
print('Comuna:',df_lesiones['comuna'].unique()) # Cambiar "No especificada" y "NaN" por "SD"
print('--------------------------------------------')
print('Longitud:',(df_lesiones['longitud'] == "SD").sum()) # veirificar si se pueden imputar datos con comuna y Geo_CABA.
print('Latitud:',(df_lesiones['latutid'] == "SD").sum())
print('--------------------------------------------')
print('Participantes:',df_lesiones['participantes'].unique()) # demasiadas Variacioanes, eliminar

Podemos observar que para la columna de "Victima" se tiene valores que nombrados de diferente fomra que el dataset de Homicidios y deben normalizarse. También que aprentemente la mayoría de las columnas no tenían datos nulos, pero si "SD" (sin dato) y que por lo general son una cantidad significativa. Así que para poder usar este dataset, es conveniente tratar de imputar al máximo los registros "SD". En caso de la gravedad del incidente, el artículo "NOTAS_LESIONES_SINIESTRO_VIAL" extraído de la misma funte de los datos, indica que "SD" para este campo es en mayor probabilidad "LEVE". Para el caso de las coordenadas, se utlizarán la georeferenciación plana para imputar algunos nulos (SD) y en su defecto la ubicación de la comuna.

#### 3.1.1. Revisar columna de "COMUNA", cambiar datos nulos y "No especificada" por "SD":

# Imputar "SD" en lugar de "0" en la columna 'comuna'
df_lesiones.loc[df_lesiones['comuna'] == 'No Especificada', 'comuna'] = 'SD'
df_lesiones.loc[df_lesiones['comuna'].isna(), 'comuna'] = 'SD'
print('Comuna:',df_lesiones['comuna'].unique())

#### 3.1.2. Normalizar datos de "Victima" y "Acusado" a los valores del dataset de "Homicidios":

Columna "Victima":

df_lesiones.loc[df_lesiones['victima'] == 'CICLISTA', 'victima'] = 'BICICLETA'
df_lesiones.loc[df_lesiones['victima'] == 'TRANSPORTE PUBLICO', 'victima'] = 'PASAJEROS'
df_lesiones.loc[df_lesiones['victima'] == 'CAMIONETA', 'victima'] = 'AUTO'
df_lesiones.loc[df_lesiones['victima'] == 'CAMION', 'victima'] = 'CARGAS'
df_lesiones.loc[df_lesiones['victima'] == 'MONOPATIN', 'victima'] = 'OTRO'
df_lesiones.loc[df_lesiones['victima'] == 'UTILITARIO', 'victima'] = 'AUTO'
df_lesiones.loc[df_lesiones['victima'] == 'TAXI', 'victima'] = 'PASAJEROS'
df_lesiones.loc[df_lesiones['victima'] == 'MIXTO', 'victima'] = 'MULTIPLE'

Columna "Acusado":

df_lesiones.loc[df_lesiones['acusado'] == 'CICLISTA', 'acusado'] = 'BICICLETA'
df_lesiones.loc[df_lesiones['acusado'] == 'TRANSPORTE PUBLICO', 'acusado'] = 'PASAJEROS'
df_lesiones.loc[df_lesiones['acusado'] == 'CAMIONETA', 'acusado'] = 'AUTO'
df_lesiones.loc[df_lesiones['acusado'] == 'CAMION', 'acusado'] = 'CARGAS'
df_lesiones.loc[df_lesiones['acusado'] == 'MONOPATIN', 'acusado'] = 'OTRO'
df_lesiones.loc[df_lesiones['acusado'] == 'UTILITARIO', 'acusado'] = 'AUTO'
df_lesiones.loc[df_lesiones['acusado'] == 'TAXI', 'acusado'] = 'PASAJEROS'
df_lesiones.loc[df_lesiones['acusado'] == 'MIXTO', 'acusado'] = 'MULTIPLE'

#### 3.1.3. Imputar datos de Coordenadas:

Crear función que busca las transforma las coordenada planas de CABA a Latitud y Longitud:

ver: https://epsg.io/transform#s_srs=9498&t_srs=4326&x=104991.4633392&y=107720.9664748

# Define la proyección original (EPSG:5343) con transformación 9264
proyeccion_origen = Proj(init='epsg:9498 ', preserve_units=False)

# Define la proyección de destino (WGS84 para latitud y longitud)
proyeccion_destino = Proj(init='epsg:4326')

# Función para convertir coordenadas planas a latitud y longitud
def convertir_a_latitud_longitud(x, y):
    try:
        latitud, longitud = transform(proyeccion_origen, proyeccion_destino, x, y)
        return latitud, longitud
    except Exception as e:
        print(f"Error: {e}")
    return None, None

Usar fórmula para convertir coordenadas planas CABA en coordenadas mundiales:


# Extraer las coordenadas X y Y de la columna 'geocodificacion_CABA'
df_lesiones[['X', 'Y']] = df_lesiones['geocodificacion_CABA'].str.extract(r'Point \((.*?) (.*?)\)').astype(float)

# Aplicar la función a las columnas de coordenadas X y Y, y crear nuevas columnas para latitud y longitud
df_lesiones[['Latitud', 'Longitud']] = df_lesiones.apply(lambda row: convertir_a_latitud_longitud(row['X'], row['Y']), axis=1, result_type='expand')

# Mostrar el DataFrame actualizado
df_lesiones

df_lesiones[['Latitud','Longitud','latutid','longitud']].head(10)

Se observan diferencias en la conversión debido a que no se conocía con la precisión ajustaremos con el error medio:

# Filtra los registros con valores menores a -30 en la columna 'Latitud'
df_filtrado = df_lesiones[(df_lesiones['Latitud'] < -38) & (df_lesiones['Latitud'] > -60) & (df_lesiones['latutid'] > -38)]


# Calcula los promedios excluyendo los valores nulos
promedio_latitud = df_filtrado['Latitud'].mean(skipna=True)
promedio_longitud = df_filtrado['Longitud'].mean(skipna=True)
promedio_latutid = df_filtrado['latutid'].mean(skipna=True)
promedio_longitud_2 = df_filtrado['longitud'].mean(skipna=True)

# Calcula las diferencias promedio
dif_Lat = promedio_latitud - promedio_longitud_2
dif_Lon = promedio_longitud- promedio_latutid

print(f"Diferencia entre promedio de Latitud y promedio de Longitud: {dif_Lat}")
print(f"Diferencia entre promedio de Longitud y promedio de latutid: {dif_Lon}")

Conociendo el error medio corregimos los datos convertidos de coordenadas:

df_lesiones['Latitud'] = df_lesiones['Latitud'] - dif_Lat
df_lesiones['Longitud'] = df_lesiones['Longitud'] - dif_Lon

df_lesiones[['Latitud','Longitud','latutid','longitud']].head(5)

Verificamos que los registros se han corregido.

Imputar los datos de coordenadas con los datos extraídos de las coordenadas planas:

# Imputar los datos vacíos y aquellos con "SD" en "latutid" con los valores de "Longitud"
df_lesiones['latutid'].fillna(df_lesiones['Longitud'], inplace=True)
df_lesiones['latutid'] = np.where(df_lesiones['latutid'] == 'SD', df_lesiones['Longitud'], df_lesiones['latutid'])

# Imputar los datos vacíos y aquellos con "SD" en "longitud" con los valores de "Latitud"
df_lesiones['longitud'].fillna(df_lesiones['Latitud'], inplace=True)
df_lesiones['longitud'] = np.where(df_lesiones['Longitud'] == 'SD', df_lesiones['Latitud'], df_lesiones['longitud'])

# Renombrar 'latutid' con 'latitud'
df_lesiones.rename(columns={'latutid': 'latitud'}, inplace=True)

# Mostrar el DataFrame actualizado
df_lesiones.head()

### 3.2. Elimnar las columnas innecesarias:

df_lesiones.columns

# Lista de columnas a eliminar
columnas_lec_eliminar = ['n_victimas', 'aaaa', 'mm', 'dd', 'franja_hora',
       'direccion_normalizada', 'otra_direccion',
       'calle', 'altura', 'cruce', 'geocodificacion_CABA', 'participantes', 'moto', 'auto',
       'transporte_publico', 'camion', 'ciclista', 'gravedad', 'X', 'Y',
       'Latitud', 'Longitud']

# Eliminar las columnas especificadas
df_lesiones = df_lesiones.drop(columns=columnas_lec_eliminar)

# Formato de fecha
df_lesiones['fecha'] = pd.to_datetime(df_lesiones['fecha'], errors='coerce')

# Mostrar el DataFrame actualizado
df_lesiones

Se concidera liasta esta tabla para trabajaren el EDA.

df_lesiones.info()

### 3.3. Exportar datos en CSV a la carpeta "Datos_Limpios"

df_lesiones.to_csv('Datos_limpios/Hechos_Lesiones_CABA.csv',index=False)

#### 4. Crear Datafrome de la hoja de "Victimas" del archivo de excel "Lesiones":

# Crear un objeto ExcelFile para manejar el archivo Excel
xl = pd.ExcelFile(Ubicación_lesiones)

# crear dataframe
df_Lesionados = xl.parse('VICTIMAS',skiprows=0)

# Reiniciar los índices después de la eliminación
df_Lesionados.reset_index(drop=True, inplace=True)

#### 4.1. Verificamos la calidad de los datos

df_Lesionados.info()

Aparente mente no se encuentran valores nulos sin embargo se precede a revisar si existe registros como "SD" como en el dataset anterior:

print('Victima:',df_Lesionados['VEHICULO_VICTIMA'].unique())  # Se debe normalizar los datos en función del archivo de Homicidios
print('--------------------------------------------')   
print('sexo:',df_Lesionados['SEXO'].unique())   # Se debe transformar valores en mayúsculas
print('--------------------------------------------')
print('Edad:',df_Lesionados['EDAD_VICTIMA'].unique())  # Se deb transformar valores en mayúsculas
print('--------------------------------------------')
print('Gravedad:',df_Lesionados['GRAVEDAD'].unique())  # Se deb transformar valores en mayúsculas

#### 4.2. Realizar transformaciones de valores en minúscula y "SD" a "leve" en la columna gravedad:

# Cambiar a mayúsculas
df_Lesionados['VEHICULO_VICTIMA'] = df_Lesionados['VEHICULO_VICTIMA'].str.upper()
df_Lesionados['SEXO'] = df_Lesionados['SEXO'].str.upper()
df_Lesionados['GRAVEDAD'] = df_Lesionados['GRAVEDAD'].str.upper()

# Reemplazar "SD" por "LEVE" en la columna "GRAVEDAD"
df_Lesionados['GRAVEDAD'] = df_Lesionados['GRAVEDAD'].replace("SD", "LEVE") # Esto basado en la consideración escrita en "NOTAS_HOMICIDIOS..."

print('Victima:',df_Lesionados['VEHICULO_VICTIMA'].unique()) 
print('--------------------------------------------')   
print('sexo:',df_Lesionados['SEXO'].unique())  
print('--------------------------------------------')
print('Edad:',df_Lesionados['EDAD_VICTIMA'].unique()) 
print('--------------------------------------------')
print('Gravedad:',df_Lesionados['GRAVEDAD'].unique())


#### 4.2.1. Cambiar valores con errores y normalizar los datos en función de los valores en el dataset de homicidios:

df_Lesionados['EDAD_VICTIMA'] = df_Lesionados['EDAD_VICTIMA'].replace("sd", "SD")
df_Lesionados['SEXO'] = df_Lesionados['SEXO'].replace("MUJER ", "MUJER")

# Normalizamos valores de "Victima" según los datos del dataset de homicidios:

df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'CICLISTA', 'VEHICULO_VICTIMA'] = 'BICICLETA'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'TRANSPORTE PUBLICO', 'VEHICULO_VICTIMA'] = 'PASAJEROS'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'CAMIONETA', 'VEHICULO_VICTIMA'] = 'AUTO'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'CAMION', 'VEHICULO_VICTIMA'] = 'CARGAS'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'MONOPATIN', 'VEHICULO_VICTIMA'] = 'OTRO'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'UTILITARIO', 'VEHICULO_VICTIMA'] = 'AUTO'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'TAXI', 'VEHICULO_VICTIMA'] = 'PASAJEROS'
df_Lesionados.loc[df_Lesionados['VEHICULO_VICTIMA'] == 'MIXTO', 'VEHICULO_VICTIMA'] = 'MULTIPLE'

Lo anterior se hace para que al comparar con el datset de homicidios, se cuente con los mismos valores.

print('Victima:',df_Lesionados['VEHICULO_VICTIMA'].unique()) 
print('--------------------------------------------')   
print('sexo:',df_Lesionados['SEXO'].unique())  
print('--------------------------------------------')
print('Edad:',df_Lesionados['EDAD_VICTIMA'].unique()) 
print('--------------------------------------------')
print('Gravedad:',df_Lesionados['GRAVEDAD'].unique())


df_Lesionados.columns

#### 4.3. Eliminar columnas innecesarias y cambiar formato de tipos de dato correspondiente:

# Lista de columnas a eliminar
columnas_lec_vic_eliminar = ['AAA', 'MM', 'DD']

# Eliminar las columnas especificadas
df_Lesionados = df_Lesionados.drop(columns=columnas_lec_vic_eliminar)

# Formato de fecha
df_Lesionados['EDAD_VICTIMA'] = pd.to_numeric(df_Lesionados['EDAD_VICTIMA'], errors='coerce')

# Mostrar el DataFrame actualizado
df_Lesionados

#### 4.4. Exportar dataframe a CSV en la carpeta "Datos_Limpios"

df_Lesionados.to_csv('Datos_limpios/Victimas_Lesiones_CABA.csv',index=False)
