==> 4. ETL_General.ipynb <==
# ETL (Extracción, Transformación y Limpieza General)

Al revisar los datos limpios, se observa que se deben corregir algunos datos, al tratarlos de abrir en power bi, se determina que es más efectivo cambiar la estructura de la información evitando las uniones de datos en la herramienta de visualización.

A continuación se siguen analizando la estructura de los datos y transformaciones necesarias:

#### 1. Importar librerías:

import pandas as pd                                                           # Para abrir y transformar los archivos
import numpy as np                                                            # Cálculos y manipulación de datos

#### 2. Abrir archivos relacionados a los siniestros viales:

H_Homicidios = pd.read_csv('Datos_limpios\Hechos_Homicidios_CABA.csv')        # Datos de Hechos de Homicidios
V_Homicidios = pd.read_csv('Datos_limpios\Victimas_Homicidios_CABA.csv')      # Datos de Víctimas de Homicidios
H_Lesiones = pd.read_csv('Datos_limpios\Hechos_Lesiones_CABA.csv')            # Datos de Hechos de Lesiones
V_Lesiones = pd.read_csv('Datos_limpios\Victimas_Lesiones_CABA.csv')          # Datos de Víctimas de Lesiones
Poblacion = pd.read_csv('Datos_limpios\Población_CABA.csv')                   # Datos de Población

# Revisar que las columnas sean iguales, para Hechos en Lesiones y Homicidios

print(H_Homicidios.columns)
print(H_Lesiones.columns)

Revisar si los datos están en el mismo periódo de análisis:

print('El archvio de Homicidios tiene registros desde ', H_Homicidios['FECHA'].min(),' hasta ', H_Homicidios['FECHA'].min())
print('El archvio de Homicidios tiene registros desde ', H_Lesiones['fecha'].min(),' hasta ', H_Lesiones['fecha'].min())

Aunque se unirán los dataframes, se tendrá presente que los análisis que involucren Lesiones deben realizarse solo para los años 2019 al 2021

#### 3. Unir los dataframe de "Hechos" para "Homicidios" y "Lesiones":

Se llega a la conclusión que es mejor manejar un archivo grande para Hechos y otro para Víctimas, esto no se realizó en el ETL porque en los db se podían gestionar las uniones.

# Agregar columna "TIPO" a cada DataFrame
H_Homicidios['TIPO'] = 'HOMICIDIO'
H_Lesiones['TIPO'] = 'LESION'

# Renombrar columnas en H_Lesiones para que coincidan con H_Homicidios
H_Lesiones = H_Lesiones.rename(columns={
    'id': 'ID',
    'fecha': 'FECHA',
    'hora': 'HORA',
    'comuna': 'COMUNA',
    'tipo_calle': 'TIPO_DE_CALLE',
    'longitud': 'pos x',
    'latitud': 'pos y',
    'victima': 'VICTIMA',
    'acusado': 'ACUSADO'
})

# Reorganizar las columnas en H_Homicidios
H_Homicidios = H_Homicidios[['ID', 'FECHA', 'HORA', 'COMUNA','TIPO_DE_CALLE', 'pos x', 'pos y', 'VICTIMA', 'ACUSADO', 'TIPO']]

# Concatenar verticalmente los DataFrames
Hechos = pd.concat([H_Homicidios, H_Lesiones], ignore_index=True)


Hechos.info()

#### 4. Arreglar formato de fecha y valores "SD" como nulos para poderlos administrar en el Dashborad:

# Reemplazar "SD" con NaN en ambos DataFrames
Hechos.replace('SD', np.nan, inplace=True)

# Convertir la columna 'FECHA' al tipo datetime
Hechos['FECHA'] = pd.to_datetime(Hechos['FECHA'], errors='coerce')

Hechos.info()

#### 4.1. corregir formato de coordenadas:

Hechos['pos y'] = Hechos['pos y'].replace("inf", np.nan)

# Corregir los valores para el dato con "ID" 24480
Hechos.loc[Hechos['ID'] == 'LC-2020-0244615', 'pos x'] = -58.29963273
Hechos.loc[Hechos['ID'] == 'LC-2020-0244615', 'pos y'] = -34.72890716

# Convertir la columna 'pos x' a texto
Hechos['pos x'] = Hechos['pos x'].astype(float)
Hechos['pos y'] = Hechos['pos y'].astype(float)

# Reemplazar valores infinitos por NaN en las columnas 'pos x' y 'pos y'
Hechos['pos x'] = Hechos['pos x'].replace([np.inf, -np.inf], np.nan)
Hechos['pos y'] = Hechos['pos y'].replace([np.inf, -np.inf], np.nan)

# Mostrar el DataFrame actualizado
Hechos

Se detectada un error en el formato de coordenadas y un dato atípico. Revisar calidad de las coordenadas a continuación:

print('pos y máxima: ',Hechos['pos y'].max())
print('pos y mínima: ',Hechos['pos y'].min())
print('pos x máxima: ',Hechos['pos x'].max())
print('pos x mínima: ',Hechos['pos x'].min())

#### 5. Exportar a CSV en la carpeta "Datos_Dashboard":

# Hechos.to_csv('Datos_Dashboard/Hechos.csv',index='False')

"""Esto ya no es necesario porque se decide unir el "Homicidios" con "Lesionados"""

#### 6. Unir las tablas de víctimas de "Lesiones" y "Homicidios":

# Revisar columnas

print('Homicidios:',V_Homicidios.columns)
print('Lesiones:',V_Lesiones.columns)

Agregar columnas faltantes:

# Agregar columna "TIPO" a cada DataFrame
V_Homicidios['TIPO'] = 'HOMICIDIO'

# Agregar Columna "ROL" a Lesiones
V_Lesiones['ROL'] = 'SD'

Renombrar columnas para que hayan coincidencias entre los dataframes y reorganizarlas:

# Renombrar columnas en V_Lesiones para que coincidan con V_Homicidios
V_Lesiones = V_Lesiones.rename(columns={
    'ID hecho': 'ID_hecho',
    'FECHA ': 'FECHA',
    'VEHICULO_VICTIMA': 'VICTIMA',
    'EDAD_VICTIMA': 'EDAD',
    'GRAVEDAD': 'TIPO',
    })

# Reorganizar las columnas en V_Homicidios
V_Homicidios = V_Homicidios[['ID_hecho', 'FECHA', 'ROL', 'VICTIMA', 'SEXO', 'EDAD', 'TIPO']]

# Concatenar verticalmente los DataFrames
Victimas = pd.concat([V_Homicidios, V_Lesiones], ignore_index=True)

Victimas

#### 7. Dar formato a las columnas e imputar valores faltantes:

# Reemplazar "SD" con NaN en ambos DataFrames
Victimas.replace('SD', np.nan, inplace=True)

# Convertir la columna 'FECHA' al tipo datetime
Victimas['FECHA'] = pd.to_datetime(Victimas['FECHA'], errors='coerce')

# Convertir la columna 'FECHA' al tipo datetime
Victimas['EDAD'] = pd.to_numeric(Victimas['EDAD'], errors='coerce')

Victimas.info()

Imputar datos para ROL:

def imputar_rol(row):
    if pd.isnull(row['ROL']) and row['VICTIMA'] == 'PASAJEROS':
        return 'PASAJERO_ACOMPAÑANTE'
    elif pd.isnull(row['ROL']) and pd.isnull(row['VICTIMA']):
        return 'SD'
    elif pd.isnull(row['ROL']) and not pd.isnull(row['VICTIMA']):
        return 'CONDUCTOR'
    else:
        return row['ROL']

# Aplicar la función a la columna 'ROL'
Victimas['ROL'] = Victimas.apply(imputar_rol, axis=1)

Recordando que este campo no existía para las Lesiones y se tratará de imputar a partir de los valores de las víctimas, por ejemplo si el vehículo es de pasajeros lo más probable es que sea un pasajero, si es peatón obviamente es peatón, etc.

Victimas['ROL'].unique()

Victimas.info()

Desafortunadamente se observa que no se registran muchos datos del "ROL" de la víctima ni del tipo de vehículo en donde se transportaba la víctima. Se presume un sistema muy débil para el registro de esta infromación. Se conservan los datos para el conteo de lesiones, pero se tiene en cuenta que discriminar por estos campos no es relevante. Sin embargo, para las víctimas que terminaron en muerte, estos campos tenían más del 98% de información, lo cual se debe tener presente.

print("Columnas en df_victimas:", Victimas.columns)
print("Columnas en df_hechos:", Hechos.columns)


#### 8. Combinar los dataframe de "Hechos" con "Vícimas":

Esto ayudará a los gráficos en el dashboard y la creación de los KPI's para medir el desempeños de las políticas públicas (planes de acción del gobierno) para disminuir el índice de mortalidad por accidentes viales:

8.1. Corrección de datos y unión de dataframes:

# Eliminar las columnas "VICTIMA" y "FECHA" del DataFrame de víctimas
df_victimas = Victimas.drop(['VICTIMA','FECHA'], axis=1)

Hechos = Hechos.drop(['TIPO'], axis=1)

# Unir los dos DataFrames utilizando el ID_hecho como clave
Siniestros_Viales = pd.merge(df_victimas, Hechos, left_on='ID_hecho', right_on='ID')

# Eliminar la columna "ID" del DataFrame de víctimas
Siniestros_Viales = Siniestros_Viales.drop(['ID'], axis=1)

# Corregir valores de "SEXO" ("Varon por Masculino y Mujer por Femenino")
Siniestros_Viales.loc[Siniestros_Viales['SEXO'] == 'VARON', 'SEXO'] = 'MASCULINO'
Siniestros_Viales.loc[Siniestros_Viales['SEXO'] == 'MUJER', 'SEXO'] = 'FEMENINO'

# Renombrar las columna de "ID_Hecho"
Siniestros_Viales = Siniestros_Viales.rename(columns={'ID_hecho': 'ID'})

# ver nuevo dataframe de "Siniestros_Viales"
print(Siniestros_Viales.info())
Siniestros_Viales


#### 9. Crear columnas auxiliares para el cálculo de KPI's:

Se decide crear columnas que faciliten las uniones con los datos población para los KPI's.

# Crear las columnas AÑO, MES y SEMESTRE
Siniestros_Viales['AÑO'] = Siniestros_Viales['FECHA'].dt.year.astype(int)
Siniestros_Viales['MES'] = Siniestros_Viales['FECHA'].dt.month
Siniestros_Viales['SEMESTRE'] = Siniestros_Viales['FECHA'].dt.month.apply(lambda x: 1 if x <= 6 else 2)

# Crear la columna Aux_Sem
Siniestros_Viales['Aux_Sem'] = Siniestros_Viales['AÑO'].astype(str) + '-' + Siniestros_Viales['COMUNA'] + '-' + Siniestros_Viales['SEMESTRE'].astype(str)

# Mostrar el DataFrame resultante
Siniestros_Viales

Siniestros_Viales['VICTIMA'].unique()

# Filtra los datos donde 'VICTIMA' es igual a "MULTIPLE"
datos_multiple = Siniestros_Viales[Siniestros_Viales['VICTIMA'] == 'MULTIPLE']

# Muestra los datos filtrados
datos_multiple

#### 10. Exportar a CSV a la carpeta "Datos_Dashboard":

Antes de exportar, durante el desarrollo del EDA, surgieron nuevos requerimientos de transformación, en especial de coordenadas, ya que en la transformación de coordenadas plana en mundiales, surgío una desviación sistemática corregida con el promedio del error promedio de los datos conocidos, lo cual se aprecia en el EDA funcionó para corregir correctamente las coordenadas. (Esto para los datos originados en Lesiones ya que en Homicidios no se tenía este problema)

# Se requiere una última modificación detectada en el EDA:

# Condiciones para reemplazar en la columna 'ROL'
condiciones = (Siniestros_Viales['VICTIMA'] == 'BICICLETA') | (Siniestros_Viales['VICTIMA'] == 'PEATON')

# Aplicar el reemplazo en la columna 'ROL' solo para las filas que cumplen las condiciones
Siniestros_Viales.loc[condiciones, 'ROL'] = Siniestros_Viales.loc[condiciones, 'VICTIMA']

# Luego, realiza la sustitución específica para 'BICICLETA' y 'PEATON'
Siniestros_Viales['ROL'] = Siniestros_Viales['ROL'].replace({'BICICLETA': 'CICLISTA', 'PEATON': 'PEATON'})

# -----------------------------------------------------------------------------------------------------------------------------------

# Diccionario de Victimas
mapeo_victimas = {'OBJETO FIJO': 'PEATON', 'PEATON_MOTO': 'PEATON', 'MULTIPLE':None}

# Aplicar el reemplazo en la columna 'VICTIMA'
Siniestros_Viales['VICTIMA'] = Siniestros_Viales['VICTIMA'].replace(mapeo_victimas)

valores_unicos = Siniestros_Viales['VICTIMA'].unique()

print(valores_unicos)

# Crear rango Etario ------------------------------------------------------------------------------------------------------------------------

# Define una función para asignar las categorías de R_Etario
def asignar_r_etario(edad):
    if 0 <= edad <= 5:
        return 'Primera Infancia'
    elif 6 <= edad <= 11:
        return 'Infancia'
    elif 12 <= edad <= 18:
        return 'Adolescencia'
    elif 19 <= edad <= 26:
        return 'Juventud'
    elif 27 <= edad <= 59:
        return 'Adultez'
    elif edad >= 60:
        return 'Vejez'
    else:
        return 'Desconocido'

# Aplica la función a la columna 'EDAD' y crea la columna 'R_Etario'
Siniestros_Viales['R_Etario'] = Siniestros_Viales['EDAD'].apply(asignar_r_etario)


# Exportar a CSV ------------------------------------------------------------------------------------------------------------------------------

# Definir el orden de las columnas
column_order = ['ID', 'FECHA', 'AÑO', 'MES', 'HORA', 'SEMESTRE', 'Aux_Sem', 'VICTIMA', 'ACUSADO', 'TIPO', 'ROL', 'SEXO', 'EDAD', 'COMUNA', 'TIPO_DE_CALLE', 'pos x', 'pos y','R_Etario' ]

# Seleccionar el orden deseado
Siniestros_Viales = Siniestros_Viales[column_order]

# Guarda el DataFrame en un archivo CSV
Siniestros_Viales.to_csv('Datos_Dashboard/Siniestros_Viales.csv', index=False)

# Muestra la información del DataFrame
Siniestros_Viales.info()

Ya se tiene el documneto listo para usar en la herramienta de visualización con todos los datos arreglados, estructura transformada, valores imputados y los tipos de datos correctos.

#### 11. Datos de Población:

Se consigue inicialmente los datos del Censo de 2010 y el 2022, se usó en la operación de ETL un prorrrateo con la variación pobalcional por comuna, para crear una base anual supuniendo un crecimiento lineal. Pero Uno de los KPI's requiere seguimiento semestral, así que se transformarán los datos para crear una base con población semestral por comuna. Luego a partir de este datafrome se traerán los datos necesarios del archivo de Siestros Viales para analizar estos datos según la Población:

#### 11.1. Traer datos ordenados y cread columnas auxiliares para transformar la tabla a semestres:

# Traer datos ordenados por Comuna y Año
Poblacion_Semestre = Poblacion.sort_values(by=['Comuna', 'Año'])

Poblacion_Semestre

Se crean los campos Población_1 y Variación:

# Crear la columna auxiliar "Poblacion_1"
Poblacion_Semestre['Poblacion_1'] = Poblacion_Semestre['Población'].shift(1)

# Crear la columna auxiliar "Variacion"
Poblacion_Semestre['Variacion'] = (Poblacion_Semestre['Poblacion_1'] - Poblacion_Semestre['Población']).div(2)

# Llenar la columna "Poblacion_1" restando la "Variacion"
Poblacion_Semestre['Poblacion_1'] = Poblacion_Semestre['Poblacion_1'].sub(Poblacion_Semestre['Variacion'], fill_value=0)

# Eliminar la columna "ID" del DataFrame de víctimas
Poblacion_Semestre = Poblacion_Semestre.drop(['Variacion'], axis=1)

# Mostrar el DataFrame resultante
Poblacion_Semestre

#### 11.2. Creamos registros duplicados y generamos datos semestrales de población:

# 1. Duplicar cada fila
Poblacion_Semestre = pd.concat([Poblacion_Semestre]*2, ignore_index=True)

# 2. Crear la columna 'Semestre' con valores 1 y 2
Poblacion_Semestre['Semestre'] = (Poblacion_Semestre.groupby('Año').cumcount() % 2) + 1

# 3. Establecer como NaN los valores de la columna 'Población' para los semestres 1
Poblacion_Semestre.loc[Poblacion_Semestre['Semestre'] == 1, 'Población'] = None

# 4. Imputar los valores NaN en la columna 'Población' para los semestres 1 con los valores de 'Poblacion_1'
Poblacion_Semestre.loc[Poblacion_Semestre['Semestre'] == 1, 'Población'] = Poblacion_Semestre['Poblacion_1']

# 5. Eliminar la columna 'Poblacion_1'
Poblacion_Semestre = Poblacion_Semestre.drop(columns=['Poblacion_1'])

Observamos que queden correctas las imputaciones por comuna y semestre:

# Filtrar los registros donde el valor en la columna "Comuna" es igual a 1
registros_comuna_1 = Poblacion_Semestre[Poblacion_Semestre['Comuna'] == 1]

# Mostrar los registros resultantes
print(registros_comuna_1)

#### 11.3. Creamos una columna auxiliar de semestre que nos servirá para unir datos agragados desde la hoja de "Siniestros":

# Crear la columna Aux_Sem
Poblacion_Semestre['Aux_Sem'] = Poblacion_Semestre['Año'
    ].astype(str) + '-' + Poblacion_Semestre['Comuna'].astype(str) + '-' + Poblacion_Semestre['Semestre'].astype(str)

# Definir el orden de las columnas
order = ['Año','Semestre', 'Aux_Sem', 'Comuna',  'Población']

# Seleccionar el orden deseado
Poblacion_Semestre = Poblacion_Semestre[order]

# Mostrar el DataFrame resultante
Poblacion_Semestre

#### 11.4. Crear las columnas de "No. Muertes" y "No. Lesiones" basado en los datos de "Homicidios":

# Agrupar y contar en Siniestros_Viales
conteo_tipos = Siniestros_Viales.groupby(['Aux_Sem', 'TIPO']).size().unstack(fill_value=0)

# Filtrar solo las columnas relevantes ('HOMICIDIO', 'GRAVE', 'LEVE')
conteo_tipos = conteo_tipos[['HOMICIDIO', 'GRAVE', 'LEVE']]

# Fusionar con Poblacion_Semestre
Poblacion_Semestre = Poblacion_Semestre.merge(conteo_tipos, how='left', left_on='Aux_Sem', right_index=True)

# Renombrar columnas
Poblacion_Semestre = Poblacion_Semestre.rename(columns={'HOMICIDIO': 'No. Muertes', 'GRAVE': 'No. Graves', 'LEVE': 'No. Leves'})

# Imputar valores nulos como cero "0"
Poblacion_Semestre[['No. Muertes', 'No. Graves', 'No. Leves']] = Poblacion_Semestre[['No. Muertes', 'No. Graves', 'No. Leves']].fillna(0)

# Mostrar el resultado
print(Poblacion_Semestre)


Siniestros_Viales['TIPO'].unique()

#### 11.5. Crear THSV (tasa de homicidios de Siniestros Viales)
(Número de homicidios en siniestros viales / Población total) * 100,000

# Ordenar el DataFrame por Año, Semestre y Comuna
Poblacion_Semestre = Poblacion_Semestre.sort_values(by=['Comuna', 'Año', 'Semestre'])

# Calcular la columna THSV
Poblacion_Semestre['THSV'] = (Poblacion_Semestre['No. Muertes'] / Poblacion_Semestre['Población']) * 100000

# Crear la columna THSV_1 usando shift para obtener el valor del registro inmediatamente anterior
Poblacion_Semestre['THSV_1'] = Poblacion_Semestre['THSV'].shift(1)

# Llenar con 0 en los casos donde no haya un valor anterior (primer semestre del primer año)
Poblacion_Semestre['THSV_1'] = Poblacion_Semestre['THSV_1'].fillna(0)


Poblacion_Semestre[Poblacion_Semestre['THSV_1'] == 0]


# Crear Columna "PR_THSV" (porcentaje de reducción de la tasa de homicidios de siniestros viales)
Poblacion_Semestre['PR_THSV'] = 100*(Poblacion_Semestre['THSV'] - Poblacion_Semestre['THSV_1'])/Poblacion_Semestre['THSV_1']

Eliminar columna auxiliar:

# Eliminar la columna "THSV_1" del DataFrame de Poblacion_Semestre

# Filtrar el DataFrame
filtro_poblacion = (Poblacion_Semestre['Año'].between(2016, 2021)) & (Poblacion_Semestre['Comuna'] == 15)# & (Poblacion_Semestre['THSV_1'] == 0)
datos_filtrados = Poblacion_Semestre[filtro_poblacion]

print(datos_filtrados[['THSV','THSV_1','Año','Semestre','Comuna']] )

#Poblacion_Semestre = Poblacion_Semestre.drop(['THSV_1'], axis=1)

#### 11.5. Crear Indice de Mortalidad de Motociclistas con respecto del año anterior:

Contamos las muertes donde la víctima se relaciona con "MOTO" en el tipo de vehículo de la víctima.

# Contar el número de registros en Siniestros_Viales que cumplen con las condiciones
count_m_moto = Siniestros_Viales.loc[
    (Siniestros_Viales['Aux_Sem'].isin(Poblacion_Semestre['Aux_Sem'])) &
    (Siniestros_Viales['TIPO'] == 'HOMICIDIO') &
    (Siniestros_Viales['VICTIMA'] == 'MOTO')
].groupby('Aux_Sem')['TIPO'].count().reset_index()

# Renombrar la columna y fusionar con Poblacion_Semestre
count_m_moto = count_m_moto.rename(columns={'TIPO': 'M. Moto'})
Poblacion_Semestre = pd.merge(Poblacion_Semestre, count_m_moto, on='Aux_Sem', how='left')

# Rellenar NaN con 0 si no hay coincidencias
Poblacion_Semestre['M. Moto'] = Poblacion_Semestre['M. Moto'].fillna(0).astype(int)

# Mostrar el DataFrame resultante
Poblacion_Semestre

# Ordenar el DataFrame por Año, Semestre, Comuna y moto
Poblacion_Semestre = Poblacion_Semestre.sort_values(by=['Comuna', 'Año', 'Semestre'])

# Crear la columna THSV_1 usando shift para obtener el valor del registro inmediatamente anterior
Poblacion_Semestre['M. Moto_1'] = Poblacion_Semestre['M. Moto'].shift(1)

# Llenar con 0 en los casos donde no haya un valor anterior (primer semestre del primer año)
Poblacion_Semestre['M. Moto_1'] = Poblacion_Semestre['M. Moto_1'].fillna(0)

# Crear Columna "I. Moto"
Poblacion_Semestre['I. MOTO'] = 100 * (Poblacion_Semestre['M. Moto'] - Poblacion_Semestre['M. Moto_1']) / Poblacion_Semestre['M. Moto_1']

# Suponiendo que 'df' es tu DataFrame
df_filtrado = Poblacion_Semestre.loc[Poblacion_Semestre['Año'] == 2016]

# Mostrar el DataFrame filtrado
df_filtrado.head()

#### 11.6. Eliminar las filas innecesarias:

# Eliminar las filas correspondientes al año 2022
Poblacion_Semestre = Poblacion_Semestre.loc[Poblacion_Semestre['Año'] != 2022]
Poblacion_Semestre = Poblacion_Semestre.query('Año > 2015')

# Imputar valores infinitos con cero en las columnas PR_THSV e I. MOTO
Poblacion_Semestre['PR_THSV'] = Poblacion_Semestre['PR_THSV'].replace([np.inf, -np.inf], 0)
Poblacion_Semestre['I. MOTO'] = Poblacion_Semestre['I. MOTO'].replace([np.inf, -np.inf], 0)

Poblacion_Semestre

Para los análisis se debe tener en cuenta que la base de datos de lesiones está desde el año 2019 mientras que la base de datos de muertes desde el 2016.

Relacionar el número de muertes y lesiones por año y por comuna en el dataframe de Población:

#### 11.7. Exportar a CSV en la carpeta de "Datos_Dashboard"

Poblacion_Semestre.to_csv('Datos_Dashboard/Poblacion.csv',index='False')
